fix home site generation

+++ get_pages.sh
@@ -1,30 +1,34 @@
-#!/usr/bin/env bash
 set -e
 
-for pageName in dstats NetCDF-D cblas PydMagic clFFT-D libcerf OpenMPI parallel_algorithm; do
+for pageName in . dstats NetCDF-D cblas PydMagic clFFT-D libcerf OpenMPI parallel_algorithm; do
     echo -e "****************\nbuilding: $pageName\n****************"
-    if [ ! -d repos/${pageName}/site ]
+    if [ $pageName != . ]
     then
-        echo No site directory found for repo $pageName
-        continue
+        if [ ! -d repos/${pageName}/site ]
+        then
+            echo No site directory found for repo $pageName
+            continue
+        fi
+        rm -rf ${pageName}
+        mkdir ${pageName}
+        
+        pushd repos/${pageName}
+        ./gen_docs
+        popd
+        
+        if [ -f repos/${pageName}/site/readme_as_index ]
+        then
+            cat repos/${pageName}/README.md | tr -d '\r' > ${pageName}/index.md
+        fi
+        cp -r repos/${pageName}/site/* ${pageName}/
     fi
-    rm -rf ${pageName}
-    mkdir ${pageName}
-
-    pushd repos/${pageName}
-    ./gen_docs
-    popd
-
-    if [ -f repos/${pageName}/site/readme_as_index ]
-    then
-        cat repos/${pageName}/README.md | tr -d '\r' > ${pageName}/index.md
-    fi
-    cp -r repos/${pageName}/site/* ${pageName}/
-    
     for page in $(find ${pageName} -name \*.md); do
         page_html=${page%.*}.html
         python3 md_to_html.py $page > $page_html
-        rm $page
+        if [$pageName != .]
+        then
+            rm $page
+        fi
         if [ "$(head -n1 $page_html | tr -d '\n')" != '---' ]
         then
             sed '1i\
@@ -35,4 +39,5 @@ layout: default\
             mv _tmp_ $page_html
         fi
     done
+    break
 done
